+++ Social_Library/Social_Library-war/src/java/TransferObjectInterface/LibraryDAO.java
@@ -14,7 +14,7 @@ import TransferObject.Library;
 public interface LibraryDAO {
 
      public void createLibrary(Library library);
-     public Library readLibrary(int id);
+     public Library readLibrary(long id);
      public void updateLibrary(Library libraryOld, Library libraryNew);
      public void deleteLibrary(Library library);
